requirements.txt
aws-cdk.core>=1.133.0
aws-cdk.assertions>=1.133.0
aws-cdk.aws-sqs>=1.133.0

pytest==6.2.5
